+++ src/Session/SessionManager.cc
@@ -484,6 +484,14 @@ void SessionManager::OnMessage(WSType* ws, std::string_view sv_message, uWS::OpC
                     }
                     break;
                 }
+                case CARTA::EventType::SET_VECTOR_OVERLAY_PARAMETERS: {
+                    CARTA::SetVectorOverlayParameters message;
+                    if (message.ParseFromArray(event_buf, event_length)) {
+                        tsk = new GeneralMessageTask<CARTA::SetVectorOverlayParameters>(session, message, head.request_id);
+                        message_parsed = true;
+                    }
+                    break;
+                }
                 default: {
                     spdlog::warn("Bad event type {}!", event_type);
                     break;